refactor: move ipc logic to extension host helper process
