+++ packages/node/src/parts/CommandNotFoundError/CommandNotFoundError.js
@@ -1,9 +1,7 @@
-import * as ErrorCodes from '../ErrorCodes/ErrorCodes.js'
-
 export class CommandNotFoundError extends Error {
   constructor(id) {
     super(`Command ${id} not found`)
     this.name = 'CommandNotFoundError'
-    this.code = ErrorCodes.E_COMMAND_NOT_FOUND
+    // this.code = ErrorCodes.E_COMMAND_NOT_FOUND
   }
 }
