--- packages/node/src/parts/ErrorCodes/ErrorCodes.js
@@ -1,15 +1,0 @@
-export const ENOENT = 'ENOENT'
-
-export const EXDEV = 'EXDEV'
-
-export const EEXIST = 'EEXIST'
-
-export const ENOTDIR = 'ENOTDIR'
-
-export const E_MANIFEST_NOT_FOUND = 'E_MANIFEST_NOT_FOUND'
-
-export const BABEL_PARSER_SYNTAX_ERROR = 'BABEL_PARSER_SYNTAX_ERROR'
-
-export const E_COMMAND_NOT_FOUND = 'E_COMMAND_NOT_FOUND'
-
-export const E_NO_PROVIDER_FOUND = 'E_NO_PROVIDER_FOUND'
