--- packages/node/src/parts/ErrorType/ErrorType.js
@@ -1,4 +1,0 @@
-export const DomException = 'DOMException'
-export const ReferenceError = 'ReferenceError'
-export const SyntaxError = 'SyntaxError'
-export const TypeError = 'TypeError'
