--- packages/node/src/parts/GetResponse/GetResponse.js
@@ -1,12 +1,0 @@
-import * as Command from '../Command/Command.js'
-import * as GetErrorResponse from '../GetErrorResponse/GetErrorResponse.js'
-import * as GetSuccessResponse from '../GetSuccessResponse/GetSuccessResponse.js'
-
-export const getResponse = async (message) => {
-  try {
-    const result = await Command.execute(message.method, ...message.params)
-    return GetSuccessResponse.getSuccessResponse(message, result)
-  } catch (error) {
-    return GetErrorResponse.getErrorResponse(message, error)
-  }
-}
