--- packages/node/src/parts/GetSuccessResponse/GetSuccessResponse.js
@@ -1,9 +1,0 @@
-import * as JsonRpcVersion from '../JsonRpcVersion/JsonRpcVersion.js'
-
-export const getSuccessResponse = (message, result) => {
-  return {
-    jsonrpc: JsonRpcVersion.Two,
-    id: message.id,
-    result,
-  }
-}
