+++ packages/node/src/parts/Id/Id.js
@@ -1,5 +1,5 @@
 export const state = {
-  id: 0,
+  id: 1,
 }
 
 export const create = () => {
