--- packages/node/src/parts/JsonRpcErrorCode/JsonRpcErrorCode.js
@@ -1,1 +1,0 @@
-export const MethodNotFound = -32601
