--- packages/node/src/parts/JsonRpcVersion/JsonRpcVersion.js
@@ -1,1 +1,0 @@
-export const Two = '2.0'
