--- packages/node/src/parts/RestoreJsonRpcError/RestoreJsonRpcError.js
@@ -1,50 +1,0 @@
-import * as GetErrorConstructor from '../GetErrorConstructor/GetErrorConstructor.js'
-import { JsonRpcError } from '../JsonRpcError/JsonRpcError.js'
-import * as JsonRpcErrorCode from '../JsonRpcErrorCode/JsonRpcErrorCode.js'
-
-const constructError = (message, type, name) => {
-  const ErrorConstructor = GetErrorConstructor.getErrorConstructor(message, type)
-  if (ErrorConstructor === DOMException && name) {
-    return new ErrorConstructor(message, name)
-  }
-  if (ErrorConstructor === Error) {
-    return new Error(message)
-  }
-  return new ErrorConstructor(message)
-}
-
-export const restoreJsonRpcError = (error) => {
-  if (error && error instanceof Error) {
-    return error
-  }
-  if (error && error.code && error.code === JsonRpcErrorCode.MethodNotFound) {
-    const restoredError = new JsonRpcError(error.message)
-    restoredError.stack = error.stack
-    return restoredError
-  }
-  if (error && error.message) {
-    const restoredError = constructError(error.message, error.type, error.name)
-    if (error.data) {
-      if (error.data.stack) {
-        restoredError.stack = error.data.stack
-
-        if (error.data.codeFrame) {
-          // @ts-ignore
-          restoredError.codeFrame = error.data.codeFrame
-        }
-      }
-    } else if (restoredError.stack) {
-      // TODO accessing stack might be slow
-      const lowerStack = restoredError.stack
-      // @ts-ignore
-      const indexNewLine = lowerStack.indexOf('\n')
-      // @ts-ignore
-      restoredError.stack = error.stack + lowerStack.slice(indexNewLine)
-    }
-    return restoredError
-  }
-  if (typeof error === 'string') {
-    return new Error(`JsonRpc Error: ${error}`)
-  }
-  return new Error(`JsonRpc Error: ${error}`)
-}
